Add WIP test for navigation
