+++ src/mobile_app/test/account_setup/widgets/nick_name_form_screen_test.dart
@@ -5,6 +5,7 @@ import 'package:formz/formz.dart';
 import 'package:mobile_app/account_setup/blocs/sign_up/sign_up_bloc.dart';
 import 'package:mobile_app/account_setup/models/models.dart';
 import 'package:mobile_app/account_setup/models/name_field.dart';
+import 'package:mobile_app/account_setup/widgets/name_form_screen.dart';
 import 'package:mobile_app/account_setup/widgets/nick_name_form_screen.dart';
 import 'package:mobile_app/positive_affirmations_keys.dart';
 import 'package:mocktail/mocktail.dart';
@@ -19,6 +20,8 @@ class MockSignUpBloc extends MockBloc<SignUpEvent, SignUpState>
     implements SignUpBloc {}
 
 class MockNickNameField extends Mock implements NameField {}
+
+class MockNavigatorObserver extends Mock implements NavigatorObserver {}
 
 void main() {
   const mockValidName = 'validName';
@@ -65,6 +68,7 @@ void main() {
     });
 
     testWidgets('components are rendered', (tester) async {
+      when(() => signUpBloc.state).thenReturn(mockValidSignUpState);
       await tester.pumpWidget(NickNameFormFixture(signUpBloc));
 
       expect(
@@ -101,6 +105,48 @@ void main() {
             .errorText,
         isNull,
       );
+    });
+
+    testWidgets('pressing back button pops back to name form', (tester) async {
+      final mockObserver = MockNavigatorObserver();
+      final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
+
+      when(() => signUpBloc.state).thenReturn(const SignUpState(
+        name: const NameField.dirty(mockValidName),
+        nameStatus: FormzStatus.submissionSuccess,
+      ));
+
+      await tester.pumpWidget(MaterialApp(
+        home: NameFormScreen(),
+        // navigatorObservers: [mockObserver],
+        navigatorKey: navigatorKey,
+        onGenerateRoute: (_) => NameFormScreen.route(),
+      ));
+      print(navigatorKey.currentWidget!.key.toString());
+
+      await tester.enterText(
+        find.byKey(PositiveAffirmationsKeys.nameField),
+        mockValidName,
+      );
+
+      await tester.tap(find.byKey(PositiveAffirmationsKeys.nameSubmitButton));
+
+      await tester.pumpAndSettle();
+
+      // verify(() => mockObserver.didPush(
+      //       NickNameFormScreen.route(signUpBloc),
+      //       any(),
+      //     ));
+
+
+      // await tester.tap(find.byKey(PositiveAffirmationsKeys.changeNameButton));
+      //
+      // await tester.pumpAndSettle();
+
+      // verify(() => mockObserver.didPop(
+      //       NickNameFormScreen.route(signUpBloc),
+      //       NameFormScreen.route(),
+      //     ));
     });
 
     group('[FormWiredToBloc]', () {
